refactor(components): split group ownership
