+++ ShipCoreFramework/src/Data/Scripts/ShipCoreFramework/Client/Components/GroupComponent.Ownership.cs
@@ -1,0 +1,13 @@
+namespace ShipCoreFramework
+{
+    internal partial class GroupComponent
+    {
+        private long GetObservedOwnerId()
+        {
+            if (_runtimeStateReceived) return _runtimeOwnerId;
+
+            var ownerId = ResolveLocalOwnerId();
+            return IsIgnoredNpcGroup() ? 0 : ownerId;
+        }
+    }
+}
